feat(theme-builder): move core types to a dedicated file
and don't re-export types from type-fest
