+++ packages/@poupe-theme-builder/src/core/index.ts
@@ -10,7 +10,7 @@ import {
 
 import {
   type Color,
-} from './colors';
+} from './types';
 
 // re-export
 //
@@ -26,14 +26,11 @@ export {
 export * from './colors';
 export * from './formatter';
 export * from './mix';
+export * from './types';
 
 // DynamicScheme
 //
 export type standardDynamicSchemeFactory = (primary: Color, isDark: boolean, contrastLevel: number) => DynamicScheme;
-
-// types
-//
-export type ColorMap<K extends string> = Record<K, Hct>;
 
 // tools
 //
