+++ packages/@poupe-theme-builder/src/core/utils.ts
@@ -1,9 +1,4 @@
 /* re-export */
-export type {
-  KebabCase,
-  Simplify,
-} from 'type-fest';
-
 export {
   kebabCase,
   keys,
